+++ pstatus-notifications-workflow-ktor/src/main/kotlin/workflow/digestcounts/Counts.kt
@@ -1,0 +1,8 @@
+package gov.cdc.ocio.processingnotifications.workflow.digestcounts
+
+data class Counts(
+    val started: Int,
+    val completed: Int,
+    val failedDelivery: Int,
+    val delivered: Int
+)
