+++ pstatus-notifications-workflow-ktor/src/main/kotlin/workflow/digestcounts/UploadDigestCounts.kt
@@ -1,31 +1,21 @@
 package gov.cdc.ocio.processingnotifications.workflow.digestcounts
 
 
-data class Counts(
-    val started: Int,
-    val completed: Int,
-    val failedDelivery: Int,
-    val delivered: Int
-)
-
+// Map jurisdiction to counts
 typealias CountsByJurisdiction = Map<String, Counts>
 
+// Map data stream route to jurisdiction
 typealias CountsByDataStreamRoute = Map<String, CountsByJurisdiction>
 
+// Map data stream id to data stream route
 typealias CountsByDataStreamId = Map<String, CountsByDataStreamRoute>
 
-data class UploadMetrics(
-    val minDelta: Long,
-    val maxDelta: Long,
-    val meanDelta: Float,
-    val medianDelta: Float,
-    val minFileSize: Long,
-    val maxFileSize: Long,
-    val meanFileSize: Float,
-    val medianFileSize: Float
-)
-
+/**
+ * Digest of counts by data stream, grouped by data stream id, route and jurisdiction.
+ *
+ * @property digest Map<String, Map<String, Map<String, Counts>>>
+ * @constructor
+ */
 data class UploadDigestCounts(
     val digest: CountsByDataStreamId
 )
-
